requirements.txt
streamlit>=1.35
boto3>=1.35
python-dotenv>=1.0
